fix: resolve 5 bugs in system.c, auth.c, and header.h
  - fix saveAccountToFile using pointer syntax (-> and &) on value-type
    params; add missing comma after u.id argument
  - fix getPassword skipping id field in users.txt; fscanf now reads all
    three fields (id, name, password)
  - fix createNewAcc not assigning r.id or r.userId before saving;
    r.id is derived from record count, r.userId set from u.id
  - fix phone field overflow by changing int to long in Record struct
    and updating all format strings to %ld
  - fix getAccountFromFile reading name into a separate param instead of
    r->name; removed redundant char name[] parameter and updated callers

+++ atm-system/src/header.h
@@ -14,7 +14,7 @@ struct Record
     int userId;
     char name[100];
     char country[100];
-    int phone;
+    long phone;
     char accountType[10];
     int accountNbr;
     double amount;
